Appels retrofit ItemsDetailActivity

+++ inquirio-android/app/src/main/java/ca/obrassard/inquirio/transfer/LostItemSummary.java
@@ -1,7 +1,7 @@
 package ca.obrassard.inquirio.transfer;
 
 public class LostItemSummary {
-    public int itemID;
+    public long itemID;
     public String itemName;
     public String locationName;
     public boolean found;
